Move TrainJob marker defaulting and validation integration tests to test/integration/webhooks pkg
